changed nitro/gpp bar calculation from int to float making the reduction continuous rather than discrete
git-svn-id: svn://fly.thruhere.net/forsaken@2134 1361c5d3-c82f-4d94-b015-8aedb5f389d9

+++ ProjectX/trunk/src/oct2.c
@@ -1556,7 +1556,7 @@ void DrawSimplePanel()
 		// nitro bar
 		if ( ( control.turbo || Ships[WhoIAm].Object.CruiseControl == CRUISE_NITRO ) && NitroFuel )
 		{
-			int bar = ( Ships[WhoIAm].Object.Flags & SHIP_SuperNashram ) ?
+			float bar = ( Ships[WhoIAm].Object.Flags & SHIP_SuperNashram ) ?
 				(float) Ships[WhoIAm].SuperNashramTimer*0.008f :
 				(float) NitroFuel*0.04f;
 			AddScreenPolyTextScale(
